feat(sim+benchmark): fix SMS/GG emulation, add QEMU ESP32-S3 CPU benchmark
Simulator fixes:
- Fix smsplus uint32 typedef (unsigned long → unsigned int) for 64-bit
  macOS — was 8 bytes causing VDP tile corruption (vertical black bars)
- Fix SMS palette: proper RGB565 lookup + byte-swap for SDL
- Fix viewport.x/y offset for overscan border rows
- Force CONSOLE_GG for .gg ROMs missing TMR SEGA header
- Add test_screenshots.c for automated screenshot generation

QEMU ESP32-S3 benchmark (new):
- Docker image with ESP-IDF v5.5.4 + QEMU Xtensa (aarch64)
- Benchmark firmware: 7 cores × 300 frames with esp_timer
- ROM embedding via EMBED_FILES (no SPIFFS needed)
- VNC support via Xvfb + x11vnc + noVNC
- Results (ESP32-S3 @ 240MHz, 8MB PSRAM):
  SNES: 556 fps (9.3x), NES: 655 fps (10.9x), GB: 413 fps (6.9x)
  GBC: 377 fps (6.3x), SMS: 481 fps (8.0x), GG: 501 fps (8.4x)
  PCE: 625 fps (10.4x) — all cores well above 60fps target

Documentation:
- Add benchmark results table to feasibility.md
- Add 7 emulator screenshots (NES, GB, GBC, SMS, GG, PCE, SNES)

+++ software/components/smsplus/shared.h
@@ -3,11 +3,11 @@
 
 typedef unsigned char uint8;
 typedef unsigned short int uint16;
-typedef unsigned long int uint32;
+typedef unsigned int uint32;  /* must be 4 bytes — unsigned long is 8 on 64-bit */
 
 typedef signed char int8;
 typedef signed short int int16;
-typedef signed long int int32;
+typedef signed int int32;  /* must be 4 bytes */
 
 #include <assert.h>
 #include <stdio.h>
